Add ability to record and validate output of demo files
Over time the RxPHP project has added demos for various observables and
operators. From time to time the demos tend to break because of
refactoring or other changes.

This commit introduces a "test" command inside the demo folder that will
find demos that have an accompanying `.expect` file and ensure that
running the demo yields the same output as is expected. This will ensure
that demos keep working, but it also turns each demo into an end to end
tests for the RxPHP project.

Example usage:

```
$ php demo/test.php
Found 2 demos with an expect file.
[v] demo/combineLatest/combineLatest.php
[v] demo/concat/concat.php
```