Add .expect files for most demos
For most demos it is easy to add an expect file. Some demos have been
changed slightly to not run forever or run a bit faster.

There are a few demos that are time based. They might turn out to be
flaky "tests", but let's give it a shot for now as we're talking about
hundreds of milliseconds timing not microseconds.