Add utility to generate documentation for reactivex.io
In order to provide up-to-date documentation on reactivex.io we'll
extract the operator documentation directly from the source code. Most
operators already have demo code including expected output in the demo
directory.

In order for an operator to get added to the documentation the docblock
of the method in `Rx\Observable` needs a few additions:

- `@observable` or `@operator` annotation. This will flag the method for
  extraction by the documentation builder.
- `@reactivex` the id of the operator in the reactivex.io repository
- `@demo` (optional) one or multiple annotations containing the path to
  a demo inside the demo directory

For example the `map()` operation:

```
/**
 * It takes a transforming function that operates on each element.
 *
 * @param callable $selector
 * @return AnonymousObservable
 *
 * @demo map/map.php
 * @operator
 * @reactivex map
 */
public function map(callable $selector) { {
    ...
}
```

Usage:

```
$ php docs/build-docs.php
Usage:  <command>
  lint      Verifies that documentation can be loaded
  reactivex Updates the reactivex documentation
```

Note: the command currently assumes a valid checkout of the reactivex
site repository under `docs/reactivex.github.io`.